Basic script output!
Formatting is specified through a record of format functions.  The
formatter is applied through an AST traversal on the Free monad.

There is still a lot of work to do, but this is progress.

Optimization can be written as a similar pass that returns another AST
instead of a bytestring.